Only use 1 store for content suggestions
